feat: Implement main window rendering and tab management features
- Added `main_window_render.rs` for rendering the main window, including tab management and search functionality.
- Introduced `main_window_search.rs` to handle search bar events and toggling visibility.
- Created `main_window_split_pane_actions.rs` for splitting panes horizontally and vertically, and closing active panes.
- Defined `main_window_tab_item.rs` to represent individual tabs with their properties and display logic.
- Developed `main_window_tab_management.rs` for managing tab insertion, removal, and terminal event subscriptions.
- Implemented `main_window_tab_switcher_logic.rs` for handling tab switching logic and visibility.
- Updated `mod.rs` to include new modules for better organization of components.
