+++ crates/kazeterm/src/components/main_window_dialog_handlers.rs
@@ -1,0 +1,144 @@
+use gpui::{AppContext, Context, Entity, Focusable, Window};
+
+use super::main_window::MainWindow;
+use crate::components::about_dialog::{AboutDialog, AboutDialogCloseEvent};
+use crate::components::close_confirm_dialog::{CloseConfirmDialog, CloseConfirmEvent};
+use crate::components::tab_rename_dialog::{TabRenameDialog, TabRenameEvent};
+
+impl MainWindow {
+  /// Show rename dialog for a tab
+  pub(crate) fn show_rename_dialog(&mut self, tab_index: usize, window: &mut Window, cx: &mut Context<Self>) {
+    // Find the tab's current display title
+    let current_title = self
+      .items
+      .iter()
+      .find(|item| item.index == tab_index)
+      .map(|item| item.display_title().to_string())
+      .unwrap_or_default();
+
+    let dialog = cx.new(|cx| TabRenameDialog::new(tab_index, &current_title, window, cx));
+
+    let subscription = cx.subscribe_in(&dialog, window, Self::on_rename_dialog_event);
+
+    // Focus the dialog
+    dialog.update(cx, |dialog: &mut TabRenameDialog, cx| {
+      dialog.focus(window, cx);
+    });
+
+    self.rename_dialog = Some(dialog);
+    self._rename_dialog_subscription = Some(subscription);
+    cx.notify();
+  }
+
+  pub(crate) fn on_rename_dialog_event(
+    &mut self,
+    _dialog: &Entity<TabRenameDialog>,
+    event: &TabRenameEvent,
+    window: &mut Window,
+    cx: &mut Context<Self>,
+  ) {
+    let tab_index = event.tab_index;
+    let new_title = event.new_title.clone();
+
+    // Find the tab and update its custom_title
+    if let Some(item) = self.items.iter_mut().find(|item| item.index == tab_index) {
+      item.custom_title = new_title;
+    }
+
+    // Close the dialog
+    self.rename_dialog = None;
+    self._rename_dialog_subscription = None;
+
+    // Refocus the terminal
+    self.refocus_active_terminal(window, cx);
+    cx.notify();
+  }
+
+  /// Show close confirmation dialog
+  pub fn show_close_confirm_dialog(&mut self, window: &mut Window, cx: &mut Context<Self>) {
+    // Don't show if already visible
+    if self.close_confirm_dialog.is_some() {
+      return;
+    }
+
+    let dialog = cx.new(|cx| CloseConfirmDialog::new(window, cx));
+    let subscription = cx.subscribe_in(&dialog, window, Self::on_close_confirm_event);
+
+    self.close_confirm_dialog = Some(dialog);
+    self._close_confirm_subscription = Some(subscription);
+    cx.notify();
+  }
+
+  pub(crate) fn on_close_confirm_event(
+    &mut self,
+    _dialog: &Entity<CloseConfirmDialog>,
+    event: &CloseConfirmEvent,
+    window: &mut Window,
+    cx: &mut Context<Self>,
+  ) {
+    match event {
+      CloseConfirmEvent::Confirm => {
+        // User confirmed, close the window
+        self.close_confirm_dialog = None;
+        self._close_confirm_subscription = None;
+        window.remove_window();
+      }
+      CloseConfirmEvent::Cancel => {
+        // User cancelled, just close the dialog
+        self.close_confirm_dialog = None;
+        self._close_confirm_subscription = None;
+
+        // Refocus the terminal
+        self.refocus_active_terminal(window, cx);
+        cx.notify();
+      }
+    }
+  }
+
+  /// Check if close confirmation dialog is currently showing
+  pub fn is_close_confirm_visible(&self) -> bool {
+    self.close_confirm_dialog.is_some()
+  }
+
+  /// Show about dialog
+  pub fn show_about_dialog(&mut self, window: &mut Window, cx: &mut Context<Self>) {
+    // Don't show if already visible
+    if self.about_dialog.is_some() {
+      return;
+    }
+
+    let dialog = cx.new(|cx| AboutDialog::new(window, cx));
+    let subscription = cx.subscribe_in(&dialog, window, Self::on_about_dialog_event);
+
+    self.about_dialog = Some(dialog);
+    self._about_dialog_subscription = Some(subscription);
+    cx.notify();
+  }
+
+  pub(crate) fn on_about_dialog_event(
+    &mut self,
+    _dialog: &Entity<AboutDialog>,
+    _event: &AboutDialogCloseEvent,
+    window: &mut Window,
+    cx: &mut Context<Self>,
+  ) {
+    // Close the dialog
+    self.about_dialog = None;
+    self._about_dialog_subscription = None;
+
+    // Refocus the terminal
+    self.refocus_active_terminal(window, cx);
+    cx.notify();
+  }
+
+  /// Helper to refocus the active terminal after closing dialogs
+  pub(crate) fn refocus_active_terminal(&mut self, window: &mut Window, cx: &mut Context<Self>) {
+    if let Some(active_ix) = self.active_tab_ix {
+      if let Some(item) = self.items.get(active_ix) {
+        if let Some(terminal) = item.split_container.get_active_terminal() {
+          window.focus(&terminal.focus_handle(cx));
+        }
+      }
+    }
+  }
+}
